sredili code, izbacili sve nepotrebno i prebacili sve u screenove, odradili font loading na bolji nacin u index.js entry componenti
