--- .npmrc
@@ -1,2 +1,0 @@
-shamefully-hoist=true
-node-linker=hoisted
