+++ App.js
@@ -1,7 +1,3 @@
  import 'expo-router/entry';
 
  
-  import { StatusBar } from 'expo-status-bar';
-
-<StatusBar style="light" />
-
