+++ app/BottomNavigation/index.js
